decouple_combine_rs_chunks: fix SC gather-reduce partition stride + fallback gather dtype (cluster step-0 failures)
Two cluster-observed failures of the rung-6 chunked combine, both root-caused
and fixed:

(1) SC-kernel path (no fallback flags): step-0 NaN at EVERY n_chunks>1
(N-independent). ragged_gather_reduce's main_kernel derived the row-partition
stride over the SLOT arrays (src/dst indices, weights) from the INPUT BUFFER:

    row_partition_size = in_hbm_ref.shape[0] // num_row_partitions

valid only under the kernel's implicit x.rows == indices.size contract. The
chunked combine passes the FULL expert-sorted buffer with a per-chunk SLICE of
the indices (x.rows = N * indices.size), so every row partition p>=1 read the
slot arrays out of bounds -- silently (disable_bounds_checks=True) -- and
scattered garbage rows/weights to garbage destinations. Fixed to derive the
stride from the slot arrays themselves:

    row_partition_size = src_indices_hbm_ref.shape[0] // num_row_partitions

(bit-identical for the un-chunked call, where the two operands have equal row
counts). The wrapper's x row-padding is likewise decoupled from the index
count (pad to max(padded_input_size - x.rows, 0), keeping 32-bit-packing row
parity) -- also identical for the standard call. A numpy emulation of the
wrapper padding + _preprocess + inner_kernel addressing is added to
tests/chunked_combine_equiv_test.py: the buggy stride violates the slot-array
bounds for EVERY chunk at EVERY N>1 (8/8 at N=8) and never at N=1; the fixed
stride is numerically exact for all N.

(2) Fallback path (ragged_gather_fallback + ragged_gather_reduce_fallback):
Mosaic compile error 'Expected the 2nd minor dimension is aligned to the tile'
(memref<8192x128xbf16> -> memref<1024x8x128xbf16>, left_fill_zero/dma_start,
_handwritten_moe_layer). ragged_gather's _fallback_implementation applied the
f32 routing weights with dtype PROMOTION (bf16 row * f32 weight -> f32), so
_ring_ragged_unsort_bwd emitted an f32 cotangent into the tokamax gmm_v2
backward; the f32 LHS sets dims.size_lhs_sublane=8, and the bf16 output
zero-fill reshape (-1, 8, 128) then violates the bf16 (16,128) tiling. The SC
kernel's contract is out.dtype == x.dtype (it unpacks bf16 -> f32, multiplies,
repacks); the fallback now matches:

    out = (out.astype(f32) * weights[:, None].astype(f32)).astype(x.dtype)

Verified locally: AOT (virtual tpu7x-512, full deepseek3-671b 256-core record
config + manbwd + decouple_combine_rs_chunks=4) reproduced the exact Mosaic
error pre-fix; both the fallback and no-fallback configs compile clean
post-fix. CPU equivalence suite (fwd+bwd, EP 4/8, full/truncated buffers,
N 1-8) still bit-exact; v4 real-RoutedMoE smoke (no SC -> exercises the fixed
fallback dtype under autodiff) still bit-exact for N=2/4.

+++ src/maxtext/kernels/ragged/ragged_gather.py
@@ -333,7 +333,16 @@ def _fallback_implementation(
   """Fallback to (non-ragged) JAX implementation for ragged gather."""
   out = x[indices]
   if has_weights:
-    out = out * weights[:, None]
+    # Match the SC kernel's dtype contract: weights are applied in float32 and
+    # the result is written back in x.dtype (the kernel unpacks bf16 -> f32,
+    # multiplies by the f32 weight, and repacks to bf16; its output buffer is
+    # x.dtype). The previous version let bf16 * f32 PROMOTE the output to f32,
+    # which leaked an f32 cotangent out of _ring_ragged_unsort_bwd into the
+    # tokamax gmm_v2 backward: an f32-LHS gmm sets size_lhs_sublane=8, and its
+    # bf16 output zero-fill then fails Mosaic tiling ("Expected the 2nd minor
+    # dimension is aligned to the tile", memref 8192x128xbf16 ->
+    # 1024x8x128xbf16 in left_fill_zero/dma_start).
+    out = (out.astype(jnp.float32) * weights[:, None].astype(jnp.float32)).astype(x.dtype)
   return out
